# schemas/forgegate.dashboard-openapi.v1.json
{
  "components": {
    "schemas": {
      "ApiAuthChallenge": {
        "additionalProperties": false,
        "properties": {
          "challenge_id": {
            "pattern": "^chal-[0-9a-f]{32}$",
            "title": "Challenge Id",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "identity_id": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Identity Id",
            "type": "string"
          },
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "nonce_base64": {
            "maxLength": 44,
            "minLength": 44,
            "title": "Nonce Base64",
            "type": "string"
          },
          "project_ids": {
            "items": {
              "pattern": "^[a-z][a-z0-9-]{1,62}$",
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Project Ids",
            "type": "array"
          },
          "role": {
            "$ref": "#/components/schemas/IdentityRole"
          },
          "schema_version": {
            "const": "forgegate.api-auth-challenge.v1",
            "default": "forgegate.api-auth-challenge.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "server_instance_id": {
            "pattern": "^api-[0-9a-f]{32}$",
            "title": "Server Instance Id",
            "type": "string"
          },
          "trust_store_id": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Trust Store Id",
            "type": "string"
          }
        },
        "required": [
          "challenge_id",
          "server_instance_id",
          "trust_store_id",
          "identity_id",
          "role",
          "project_ids",
          "nonce_base64",
          "issued_at",
          "expires_at"
        ],
        "title": "ApiAuthChallenge",
        "type": "object"
      },
      "ApiChallengeRequest": {
        "additionalProperties": false,
        "properties": {
          "identity_id": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Identity Id",
            "type": "string"
          },
          "project_ids": {
            "items": {
              "pattern": "^[a-z][a-z0-9-]{1,62}$",
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Project Ids",
            "type": "array"
          },
          "role": {
            "$ref": "#/components/schemas/IdentityRole"
          }
        },
        "required": [
          "identity_id",
          "role",
          "project_ids"
        ],
        "title": "ApiChallengeRequest",
        "type": "object"
      },
      "ApiSessionCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "challenge_id": {
            "pattern": "^chal-[0-9a-f]{32}$",
            "title": "Challenge Id",
            "type": "string"
          },
          "signature_base64": {
            "maxLength": 88,
            "minLength": 88,
            "title": "Signature Base64",
            "type": "string"
          }
        },
        "required": [
          "challenge_id",
          "signature_base64"
        ],
        "title": "ApiSessionCreateRequest",
        "type": "object"
      },
...
      "AuditEventPage": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/AuditEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "next_after_sequence": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next After Sequence"
          },
          "schema_version": {
            "const": "forgegate.audit-event-page.v1",
            "default": "forgegate.audit-event-page.v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "events",
          "has_more"
        ],
        "title": "AuditEventPage",
        "type": "object"
      },
...
      "CandidateCreateCommand": {
        "additionalProperties": false,
        "properties": {
          "commit_sha": {
            "pattern": "^[0-9a-fA-F]{7,64}$",
            "title": "Commit Sha",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "project_id": {
            "pattern": "^[a-z][a-z0-9-]{1,62}$",
            "title": "Project Id",
            "type": "string"
          },
          "release_track": {
            "default": "pull-request",
            "pattern": "^[a-z][a-z0-9_-]{1,63}$",
            "title": "Release Track",
            "type": "string"
          },
          "source_branch": {
            "default": "main",
            "maxLength": 255,
            "minLength": 1,
            "title": "Source Branch",
            "type": "string"
          },
          "version": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "version",
          "commit_sha",
          "created_at"
        ],
        "title": "CandidateCreateCommand",
        "type": "object"
      },
      "CandidateDocument": {
        "discriminator": {
          "mapping": {
            "forgegate.release-candidate.v1": "#/components/schemas/ReleaseCandidate",
            "forgegate.release-candidate.v2": "#/components/schemas/ProfileBoundReleaseCandidate"
          },
          "propertyName": "schema_version"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/ReleaseCandidate"
          },
          {
            "$ref": "#/components/schemas/ProfileBoundReleaseCandidate"
          }
        ]
      },
...
      "DashboardActivationCompleted": {
        "additionalProperties": false,
        "properties": {
          "browser_delivery": {
            "const": "PENDING_BROWSER_POLL",
            "default": "PENDING_BROWSER_POLL",
            "title": "Browser Delivery",
            "type": "string"
          },
          "principal": {
            "$ref": "#/components/schemas/DashboardPrincipal"
          },
          "schema_version": {
            "const": "forgegate.dashboard-activation-completed.v1",
            "default": "forgegate.dashboard-activation-completed.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "const": "AUTHENTICATED",
            "default": "AUTHENTICATED",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "principal"
        ],
        "title": "DashboardActivationCompleted",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IdentityRole": {
        "enum": [
          "producer",
          "operator"
        ],
        "title": "IdentityRole",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Build-time contract only. The BFF uses an opaque HttpOnly Dashboard cookie plus exact Origin and anti-CSRF checks; this document is not served as interactive documentation.",
    "summary": "Same-origin local Dashboard boundary",
    "title": "ForgeGate Dashboard BFF",
    "version": "0.1.0a1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/app/api/activations/{activation_code}/challenge": {
      "post": {
        "operationId": "issueDashboardActivationChallenge",
        "parameters": [
          {
            "in": "path",
            "name": "activation_code",
            "required": true,
            "schema": {
              "pattern": "^FG-[A-Z2-7]{5}-[A-Z2-7]{5}$",
              "title": "Activation Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiChallengeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiAuthChallenge"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue Dashboard Activation Challenge"
      }
    },
    "/app/api/activations/{activation_code}/complete": {
      "post": {
        "operationId": "completeDashboardActivation",
        "parameters": [
          {
            "in": "path",
            "name": "activation_code",
            "required": true,
            "schema": {
              "pattern": "^FG-[A-Z2-7]{5}-[A-Z2-7]{5}$",
              "title": "Activation Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiSessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardActivationCompleted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Dashboard Activation"
      }
    },
    "/app/api/audit-events": {
      "get": {
        "operationId": "listDashboardAuditEvents",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_sequence",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Sequence",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "candidate_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dashboard Audit Events"
      }
    },
    "/app/api/candidates": {
      "post": {
        "operationId": "createDashboardCandidate",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-ForgeGate-CSRF",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forgegate-Csrf"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateCreateCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateDocument"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Dashboard Candidate"
      }
    },
    "/app/api/candidates/{candidate_id}": {
      "get": {
        "operationId": "getDashboardCandidate",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateDocument"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dashboard Candidate"
      }
    },
...
    "/app/api/projects": {
      "get": {
        "operationId": "listDashboardProjects",
        "parameters": [
          {
            "in": "query",
            "name": "after_project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Project Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredProjectPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dashboard Projects"
      }
    },
    "/app/api/projects/{project_id}/candidates": {
      "get": {
        "operationId": "listDashboardCandidates",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_candidate_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Candidate Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseCandidatePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dashboard Candidates"
      }
    },
    "/app/api/session": {
      "delete": {
        "operationId": "endDashboardSession",
        "parameters": [
          {
            "in": "header",
            "name": "X-ForgeGate-CSRF",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forgegate-Csrf"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardLogoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "End Dashboard Session"
      },
      "get": {
        "operationId": "getDashboardSession",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Dashboard Session"
      }
    }
  }
}